Move AgendaItemCategory.php
