+++ database/factories/AgendaItemCategoryFactory.php
@@ -2,7 +2,7 @@
 
 namespace Database\Factories;
 
-use App\AgendaItemCategory;
+use App\Models\AgendaItemCategory;
 use Faker\Generator as Faker;
 
 $factory->define(AgendaItemCategory::class, function (Faker $faker) {
